requirements.txt
fastapi
uvicorn
httpx
python-dotenv
duckduckgo-search
